fix 386 malloc tests, detect 386 darwin breakpoint line.
R=r
DELTA=22  (4 added, 0 deleted, 18 changed)
OCL=29929
CL=29944

+++ src/lib/malloc/malloc.go
@@ -17,7 +17,7 @@ type Stats struct {
 	EnableGC	bool;
 }
 
-func Alloc(uint64) *byte
+func Alloc(uintptr) *byte
 func Free(*byte)
 func GetStats() *Stats
 func Lookup(*byte) (*byte, uintptr)
